Performance change - create a new method in Ext.lib.Event called removeAllListeners to remove all listeners for a specific element; use this in the DOM object's removeAll method instead of removing the listeners one-by-one.
Previously this meant that for each listener for each event type, the Ext.lib.Event would loop though the set of all listeners for all elements for all event types to get the index of the listener to remove. Now, it'll just loop through all listeners for all event type in Ext.lib.Event and remove those that pertain to the specified element.